requirements.txt
customtkinter
Pillow
reportlab